orchestrator/requirements.txt
requests==2.33.0
PyYAML==6.0.3
litellm==1.82.6
json-repair==0.58.6
tenacity==8.5.0
feedparser==6.0.12
croniter==1.4.1
miniflux==0.0.18
# SaaS 多租户
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.0
cryptography==43.0.0
celery[redis]==5.4.0
redis==5.1.0
